Extruder temp being read wrong on Da Vinci 1.0
Extruder temp was being read wrong on Da Vinci 1.0.
It would appear as if it was stuck at 70 C, and would not increase /
decrease (Which will cause overheat).

This was due to NUM_EXTRUDER not being set before calling #include
"pins.h".

I changed the way this function works, and everything should be fine
now.